+++ lib/web/command.go
@@ -56,6 +56,7 @@ import (
 	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/session"
 	"github.com/gravitational/teleport/lib/teleagent"
+	"github.com/gravitational/teleport/lib/web/terminal"
 )
 
 // summaryBufferCapacity is the summary buffer size in bytes. The summary buffer
@@ -337,7 +338,7 @@ type summaryRequest struct {
 func (h *Handler) computeAndSendSummary(
 	ctx context.Context,
 	req *summaryRequest,
-	ws WSConn,
+	ws terminal.WSConn,
 ) (*tokens.TokenCount, error) {
 	// Convert the map nodeId->output into a map nodeName->output
 	namedOutput := outputByName(req.hosts, req.output)
@@ -395,7 +396,7 @@ func (h *Handler) computeAndSendSummary(
 	if err != nil {
 		return nil, trace.Wrap(err)
 	}
-	stream := NewWStream(ctx, ws, log, nil)
+	stream := terminal.NewWStream(ctx, ws, log, nil)
 	_, err = stream.Write(data)
 	return tokenCount, trace.Wrap(err)
 }
@@ -573,10 +574,10 @@ type commandHandler struct {
 	sshBaseHandler
 
 	// stream is the websocket stream to the client.
-	stream *WSStream
+	stream *terminal.WSStream
 
 	// ws a raw websocket connection to the client.
-	ws WSConn
+	ws terminal.WSConn
 
 	// mfaAuthCache is a function that caches the result of a function that
 	// returns a list of ssh.AuthMethods. It is used to cache the result of
@@ -589,8 +590,8 @@ type commandHandler struct {
 }
 
 // sendError sends an error message to the client using the provided websocket.
-func (t *sshBaseHandler) sendError(errMsg string, err error, ws WSConn) {
-	envelope := &Envelope{
+func (t *sshBaseHandler) sendError(errMsg string, err error, ws terminal.WSConn) {
+	envelope := &terminal.Envelope{
 		Version: defaults.WebsocketVersion,
 		Type:    defaults.WebsocketError,
 		Payload: fmt.Sprintf("%s: %s", errMsg, err.Error()),
@@ -617,7 +618,7 @@ func (t *commandHandler) ServeHTTP(_ http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	envelope := &Envelope{
+	envelope := &terminal.Envelope{
 		Version: defaults.WebsocketVersion,
 		Type:    defaults.WebsocketSessionMetadata,
 		Payload: string(sessionMetadataResponse),
@@ -639,7 +640,7 @@ func (t *commandHandler) ServeHTTP(_ http.ResponseWriter, r *http.Request) {
 }
 
 func (t *commandHandler) handler(r *http.Request) {
-	t.stream = NewWStream(r.Context(), t.ws, t.log, nil)
+	t.stream = terminal.NewWStream(r.Context(), t.ws, t.log, nil)
 
 	// Create a Teleport client, if not able to, show the reason to the user in
 	// the terminal.
@@ -653,7 +654,7 @@ func (t *commandHandler) handler(r *http.Request) {
 	t.log.Debug("Creating websocket stream")
 
 	// Start sending ping frames through websocket to the client.
-	go startPingLoop(r.Context(), t.ws, t.keepAliveInterval, t.log, t.Close)
+	go startWSPingLoop(r.Context(), t.ws, t.keepAliveInterval, t.log, t.Close)
 
 	// Pump raw terminal in/out and audit events into the websocket.
 	t.streamOutput(r.Context(), tc)
@@ -685,7 +686,7 @@ func (t *commandHandler) streamOutput(ctx context.Context, tc *client.TeleportCl
 		return
 	}
 
-	if err := t.stream.SendCloseMessage(sessionEndEvent{NodeID: t.sessionData.ServerID}); err != nil {
+	if err := t.stream.SendCloseMessage(t.sessionData.ServerID); err != nil {
 		t.log.WithError(err).Error("Unable to send close event to web client.")
 		return
 	}
@@ -696,7 +697,7 @@ func (t *commandHandler) streamOutput(ctx context.Context, tc *client.TeleportCl
 // connectToNodeWithMFA attempts to perform the mfa ceremony and then dial the
 // host with the retrieved single use certs.
 // If called multiple times, the mfa ceremony will only be performed once.
-func (t *commandHandler) connectToNodeWithMFA(ctx context.Context, ws WSConn, tc *client.TeleportClient, accessChecker services.AccessChecker, getAgent teleagent.Getter, signer agentless.SignerCreator) (*client.NodeClient, error) {
+func (t *commandHandler) connectToNodeWithMFA(ctx context.Context, ws terminal.WSConn, tc *client.TeleportClient, accessChecker services.AccessChecker, getAgent teleagent.Getter, signer agentless.SignerCreator) (*client.NodeClient, error) {
 	authMethods, err := t.mfaAuthCache(func() ([]ssh.AuthMethod, error) {
 		// perform mfa ceremony and retrieve new certs
 		authMethods, err := t.issueSessionMFACerts(ctx, tc, t.stream)
@@ -720,7 +721,7 @@ func (t *commandHandler) Close() error {
 }
 
 // makeClient builds a *client.TeleportClient for the connection.
-func (t *commandHandler) makeClient(ctx context.Context, ws WSConn) (*client.TeleportClient, error) {
+func (t *commandHandler) makeClient(ctx context.Context, ws terminal.WSConn) (*client.TeleportClient, error) {
 	ctx, span := tracing.DefaultProvider().Tracer("command").Start(ctx, "commandHandler/makeClient")
 	defer span.End()
