+++ lib/web/terminal_test.go
@@ -41,6 +41,7 @@ import (
 	"github.com/gravitational/teleport/lib/defaults"
 	"github.com/gravitational/teleport/lib/session"
 	"github.com/gravitational/teleport/lib/utils"
+	"github.com/gravitational/teleport/lib/web/terminal"
 )
 
 // TestTerminalReadFromClosedConn verifies that Teleport recovers
@@ -57,7 +58,7 @@ func TestTerminalReadFromClosedConn(t *testing.T) {
 			t.Errorf("couldn't upgrade websocket connection: %v", err)
 		}
 
-		envelope := Envelope{
+		envelope := terminal.Envelope{
 			Type:    defaults.WebsocketRaw,
 			Payload: "hello",
 		}
@@ -76,7 +77,7 @@ func TestTerminalReadFromClosedConn(t *testing.T) {
 
 	ctx, cancel := context.WithCancel(context.Background())
 	t.Cleanup(cancel)
-	stream := NewTerminalStream(ctx, TerminalStreamConfig{WS: conn, Logger: utils.NewLoggerForTests()})
+	stream := terminal.NewStream(ctx, terminal.StreamConfig{WS: conn, Logger: utils.NewLoggerForTests()})
 
 	// close the stream before we attempt to read from it,
 	// this will produce a net.ErrClosed error on the read
@@ -86,9 +87,9 @@ func TestTerminalReadFromClosedConn(t *testing.T) {
 	require.NoError(t, err)
 }
 
-type terminal struct {
+type testTerminal struct {
 	ws     *websocket.Conn
-	stream *TerminalStream
+	stream *terminal.Stream
 
 	sessionC chan session.Session
 }
@@ -101,11 +102,11 @@ type connectConfig struct {
 	participantMode   types.SessionParticipantMode
 	keepAliveInterval time.Duration
 	mfaCeremony       func(challenge client.MFAAuthenticateChallenge) []byte
-	handlers          map[string]WSHandlerFunc
-	pingHandler       func(WSConn, string) error
+	handlers          map[string]terminal.WSHandlerFunc
+	pingHandler       func(terminal.WSConn, string) error
 }
 
-func connectToHost(ctx context.Context, cfg connectConfig) (*terminal, error) {
+func connectToHost(ctx context.Context, cfg connectConfig) (*testTerminal, error) {
 	req := TerminalRequest{
 		Server: cfg.host,
 		Login:  cfg.pack.login,
@@ -170,7 +171,7 @@ func connectToHost(ctx context.Context, cfg connectConfig) (*terminal, error) {
 		})
 	}
 
-	t := &terminal{ws: ws, sessionC: make(chan session.Session, 1)}
+	t := &testTerminal{ws: ws, sessionC: make(chan session.Session, 1)}
 
 	// If MFA is expected, it should be performed prior to creating
 	// the TerminalStream to avoid messages being handled by multiple
@@ -182,11 +183,11 @@ func connectToHost(ctx context.Context, cfg connectConfig) (*terminal, error) {
 	}
 
 	if cfg.handlers == nil {
-		cfg.handlers = map[string]WSHandlerFunc{}
+		cfg.handlers = map[string]terminal.WSHandlerFunc{}
 	}
 
 	if _, ok := cfg.handlers[defaults.WebsocketSessionMetadata]; !ok {
-		cfg.handlers[defaults.WebsocketSessionMetadata] = func(ctx context.Context, envelope Envelope) {
+		cfg.handlers[defaults.WebsocketSessionMetadata] = func(ctx context.Context, envelope terminal.Envelope) {
 			if envelope.Type != defaults.WebsocketSessionMetadata {
 				return
 			}
@@ -200,7 +201,7 @@ func connectToHost(ctx context.Context, cfg connectConfig) (*terminal, error) {
 		}
 	}
 
-	t.stream = NewTerminalStream(ctx, TerminalStreamConfig{
+	t.stream = terminal.NewStream(ctx, terminal.StreamConfig{
 		WS:       ws,
 		Logger:   utils.NewLogger(),
 		Handlers: cfg.handlers,
@@ -209,34 +210,34 @@ func connectToHost(ctx context.Context, cfg connectConfig) (*terminal, error) {
 	return t, nil
 }
 
-func (t *terminal) GetSession() session.Session {
+func (t *testTerminal) GetSession() session.Session {
 	sess := <-t.sessionC
 	t.sessionC <- sess
 
 	return sess
 }
 
-func (t *terminal) Close() error {
+func (t *testTerminal) Close() error {
 	return t.stream.Close()
 }
 
-func (t *terminal) Write(p []byte) (int, error) {
+func (t *testTerminal) Write(p []byte) (int, error) {
 	return t.stream.Write(p)
 }
 
-func (t *terminal) Read(p []byte) (int, error) {
+func (t *testTerminal) Read(p []byte) (int, error) {
 	return t.stream.Read(p)
 }
 
-func (t *terminal) SetReadDeadline(deadline time.Time) error {
-	return t.stream.ws.SetReadDeadline(deadline)
+func (t *testTerminal) SetReadDeadline(deadline time.Time) error {
+	return t.stream.SetReadDeadline(deadline)
 }
 
-func (t *terminal) SetWriteDeadline(deadline time.Time) error {
-	return t.stream.ws.SetWriteDeadline(deadline)
+func (t *testTerminal) SetWriteDeadline(deadline time.Time) error {
+	return t.stream.SetWriteDeadline(deadline)
 }
 
-func (t *terminal) performMFACeremony(ceremonyFn func(challenge client.MFAAuthenticateChallenge) []byte) error {
+func (t *testTerminal) performMFACeremony(ceremonyFn func(challenge client.MFAAuthenticateChallenge) []byte) error {
 	// Wait for websocket authn challenge event.
 	ty, raw, err := t.ws.ReadMessage()
 	if err != nil {
@@ -247,7 +248,7 @@ func (t *terminal) performMFACeremony(ceremonyFn func(challenge client.MFAAuthen
 		return trace.BadParameter("got unexpected websocket message type %d", ty)
 	}
 
-	var env Envelope
+	var env terminal.Envelope
 	if err := proto.Unmarshal(raw, &env); err != nil {
 		return trace.Wrap(err, "unmarshalling envelope")
 	}
